Michael and Arnaud - Legendary Update:
Completely finished the definition of the getWinner() function. If there is one player remaining, gameIsOver is set to true and that player is the winner. If this is false, it checks if all players except one have lost, if this is true gameIsOver is set to true and that player is the winner. If all of these are false, the function returns null. Check comments for more guidance throughout the code.